+++ src/chess/piece/LookupImpl.zig
@@ -2,8 +2,30 @@ const std = @import("std");
 
 const chess = @import("../../chess.zig");
 const Side = chess.Side;
+const Color = chess.Color;
 const Piece = chess.Piece;
 const Kind = Piece.Kind;
+
+const color_kind_lut = [2][Kind.count]Piece{
+    .{
+        .white_pawn,
+        .white_knight,
+        .white_bishop,
+        .white_rook,
+        .white_queen,
+        .white_king,
+        .none,
+    },
+    .{
+        .black_pawn,
+        .black_knight,
+        .black_bishop,
+        .black_rook,
+        .black_queen,
+        .black_king,
+        .none,
+    },
+};
 
 const side_lut = Lut(Side, ConditionImpl.side);
 const kind_lut = Lut(Kind, ConditionImpl.kind);
@@ -17,6 +39,10 @@ const orthogonal_slider_lut = Lut(bool, ConditionImpl.isOrthogonalSlider);
 const char_lut = Lut(u8, ConditionImpl.char);
 
 const ConditionImpl = @import("ConditionImpl.zig");
+
+pub fn fromColorKind(color: Color, piece_kind: Kind) Piece {
+    return color_kind_lut[color.toU1()][piece_kind.toU3()];
+}
 
 pub fn side(self: Piece) Side {
     return side_lut[self.toU4()];